requirements.txt
# This file lists the dependencies of this extension.
# Install with a command like: pip install -r requirements.txt 
lxml
pyyaml==5.1.2
requests==2.5.3
ckanapi==3.4
flake8==2.1.0
ckantoolkit>=0.0.2
